khetgpt-backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
SQLAlchemy==2.0.20
pydantic==2.3.0
Pillow==10.0.0
tensorflow==2.13.0
numpy==1.24.3
matplotlib==3.7.2
scikit-learn==1.3.0
python-dotenv==1.0.0
opencv-python==4.8.0.76
scikit-image==0.21.0